add support for regex format
