+++ vocabulary/format-assertion/src/main/java/io/github/sebastiantoepfer/jsonschema/vocabulary/formatassertion/UnknownFormat.java
@@ -21,14 +21,16 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-module io.github.sebastiantoepfer.jsonschema.vocabulary.formatassertion {
-    requires io.github.sebastiantoepfer.jsonschema;
-    requires io.github.sebastiantoepfer.jsonschema.vocabulary.spi;
-    requires io.github.sebastiantoepfer.ddd.common;
-    requires java.logging;
-    requires com.github.spotbugs.annotations;
-    requires jakarta.json;
+package io.github.sebastiantoepfer.jsonschema.vocabulary.formatassertion;
 
-    provides io.github.sebastiantoepfer.jsonschema.vocabulary.spi.LazyVocabularies
-        with io.github.sebastiantoepfer.jsonschema.vocabulary.formatassertion.FormatAssertionVocabulary;
+public record UnknownFormat(String name) implements Format {
+    @Override
+    public String name() {
+        return name;
+    }
+
+    @Override
+    public boolean applyTo(final String value) {
+        return false;
+    }
 }
